+++ attributes/k8s_proxy.rb
@@ -1,0 +1,3 @@
+default['kubernetes']['proxy']['kubeconfig']    = '/etc/kubernetes/system:kube-proxy_config.yaml'
+default['kubernetes']['proxy']['feature_gates'] = node['kubernetes']['feature_gates']
+default['kubernetes']['proxy']['proxy_mode']    = 'iptables'