+++ recipes/packages.rb
@@ -6,12 +6,10 @@
 #
 
 node['kubernetes']['node']['packages'].each do |pkg, flag|
-  if flag
-    package pkg
-  end
+  package pkg if flag
 end
 
-if node['docker']['build-it']
+if node['docker']['built-in']
   package 'linux-image-extra-virtual' if node['docker']['settings']['storage-driver'] == 'aufs'
 
   apt_repository 'docker' do