+++ test/integration/roles/kubernetes_master.rb
@@ -4,35 +4,30 @@ run_list 'recipe[kubernetes::master]'
 override_attributes(
   kubernetes: {
     cluster_name: 'example',
-    cluster_dns: '192.168.222.222',
+    cluster_dns: ['192.168.222.222'],
     cluster_cidr: '192.168.0.0/17',
     api: {
       'service_cluster_ip_range' => '192.168.128.0/17',
-      'runtime_config' => %w(batch/v2alpha1)
     },
     weave: {
       use_scope: false,
-      deploy_via: 'daemonset'
+      deploy_via: 'daemonset',
     },
     dns: { deploy_via: 'deployment' },
     token_auth: true,
     addons: {
-      kubedns: {
-        node_selector: 'evl.ms/role=system'
-      },
       coredns: {
-        node_selector: 'evl.ms/role=system',
         requests: {
-          cpu: '200m'
+          cpu: '200m',
         },
         limits: {
-          cpu: '200m'
-        }
+          cpu: '200m',
+        },
       },
       dns: {
         controller: 'coredns',
-        antiaffinity_type: 'requiredDuringSchedulingIgnoredDuringExecution'
-      }
-    }
+        antiaffinity_type: 'requiredDuringSchedulingIgnoredDuringExecution',
+      },
+    },
   }
 )