+++ test/integration/roles/kubernetes_node.rb
@@ -8,18 +8,18 @@ override_attributes(
       'log-driver' => 'json-file',
       'log-opts' => {
         'max-size' => '100m',
-        'max-file' => '5'
-      }
-    }
+        'max-file' => '5',
+      },
+    },
   },
   kubernetes: {
     cluster_name: 'example',
-    cluster_dns: '192.168.222.222',
+    cluster_dns: ['192.168.222.222'],
     token_auth: true,
     api:   { 'service_cluster_ip_range' => '192.168.128.0/17' },
     weave: {
       network: '192.168.0.0/17',
-      use_scope: false
-    }
+      use_scope: false,
+    },
   }
 )